+++ app/src/main/java/com/moviles/studentcoursessystem/data/database/entity/CourseEntity.kt
@@ -1,0 +1,16 @@
+package com.moviles.studentcoursessystem.data.database.entity
+
+import androidx.room.ColumnInfo
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+@Entity(tableName = "courses")
+data class CourseEntity(
+    @PrimaryKey(autoGenerate = true)
+    @ColumnInfo(name = "id") val id: Int = 0,
+    @ColumnInfo(name = "name") val name: String,
+    @ColumnInfo(name = "description") val description: String,
+    @ColumnInfo(name = "imageUrl") val imageUrl: String?,
+    @ColumnInfo(name = "schedule") val schedule: String,
+    @ColumnInfo(name = "professor") val professor: String,
+)
