+++ app/src/main/java/com/moviles/studentcoursessystem/data/database/entity/StudentEntity.kt
@@ -1,0 +1,28 @@
+package com.moviles.studentcoursessystem.data.database.entity
+
+import androidx.room.ColumnInfo
+import androidx.room.Entity
+import androidx.room.ForeignKey
+import androidx.room.Index
+import androidx.room.PrimaryKey
+
+@Entity(
+    tableName = "students",
+    foreignKeys = [
+        ForeignKey(
+            entity = CourseEntity::class,
+            parentColumns = ["id"],
+            childColumns = ["courseId"],
+            onDelete = ForeignKey.CASCADE
+        )
+    ],
+    indices = [Index("courseId")]
+)
+data class StudentEntity(
+    @PrimaryKey(autoGenerate = true)
+    @ColumnInfo(name = "id") val id: Int = 0,
+    @ColumnInfo(name = "name") val name: String,
+    @ColumnInfo(name = "email") val email: String,
+    @ColumnInfo(name = "phone") val phone: String,
+    @ColumnInfo(name = "courseId") val courseId: Int,
+)
