+++ app/src/main/java/com/moviles/studentcoursessystem/model/Course.kt
@@ -1,0 +1,10 @@
+package com.moviles.studentcoursessystem.model
+
+data class Course(
+    val id: Int? = null,
+    val name: String,
+    val description: String,
+    val imageUrl: String?,
+    val schedule: String,
+    val professor: String
+)
