--- app/src/main/java/com/moviles/studentcoursessystem/models/Student.kt
@@ -1,30 +1,0 @@
-package com.moviles.studentcoursessystem.models
-
-import androidx.room.Entity
-import androidx.room.ForeignKey
-import androidx.room.Index
-import androidx.room.PrimaryKey
-
-@Entity(
-    tableName = "students",
-    foreignKeys = [
-        ForeignKey(
-            entity = Course::class,
-            parentColumns = ["id"],
-            childColumns = ["courseId"],
-            onDelete = ForeignKey.CASCADE
-        )
-    ],
-    indices = [Index("courseId")]
-)
-
-data class Student(
-    @PrimaryKey(autoGenerate = true)
-    val id: Int? = null,
-    val name: String,
-    val email: String,
-    val phone: String,
-    val courseId: Int,
-    val isFromCache: Boolean = false,
-    val lastSyncTimestamp: Long = System.currentTimeMillis()
-)
